requirements.txt
langchain
langchain-community
langchain-aws
langchain-pinecone
python-dotenv
pypdf
boto3
streamlit